application/requirements.txt
Fabric
Flask-WTF
WTForms
coverage
# ecdsa
# paramiko
# pycrypto
